requirements.txt
aniso8601==3.0.2
atomicwrites==1.2.1
attrs==18.2.0
colorama==0.4.1
graphene==2.1.3
graphene-tornado==2.0.1
graphql-core==2.1
graphql-relay==0.4.5
Jinja2==2.9.6
MarkupSafe==1.1.0
more-itertools==4.3.0
pluggy==0.8.0
promise==2.2.1
py==1.7.0
pytest==4.0.1
Rx==1.6.1
six==1.11.0
tornado==5.1.1
Werkzeug==0.12.2
